finished adding read time to post
